# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs for Shop Scrape App",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Retrieve regular users",
                "description": "Fetches a list of all regular users with their associated retailers.",
                "operationId": "c457726701591d1183b53aa71fc13441",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/UserResource"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Store a new user",
                "description": "Creates a new user and saves it in the database.",
                "operationId": "592819a0265360b2014512d6dbfaf0e7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Name of the user",
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "description": "Email address of the user",
                                        "type": "string",
                                        "example": "johndoe@example.com"
                                    },
                                    "password": {
                                        "description": "Password for the user",
                                        "type": "string",
                                        "example": "securepassword"
                                    },
                                    "role": {
                                        "description": "Role of the user",
                                        "type": "string",
                                        "example": "REGULAR_USER"
                                    },
                                    "location": {
                                        "description": "Location of the user",
                                        "type": "string",
                                        "example": "New York"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/users/{user}": {
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Update a user",
                "description": "Updates an existing user by its ID.",
                "operationId": "e29fc231c7937ab09483d35e6e3d1214",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID of the user to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Updated name of the user",
                                        "type": "string",
                                        "example": "Jane Doe"
                                    },
                                    "email": {
                                        "description": "Updated email address of the user",
                                        "type": "string",
                                        "example": "janedoe@example.com"
                                    },
                                    "password": {
                                        "description": "Updated password",
                                        "type": "string",
                                        "example": "newsecurepassword"
                                    },
                                    "role": {
                                        "description": "Updated role of the user",
                                        "type": "string",
                                        "example": "REGULAR_USER"
                                    },
                                    "location": {
                                        "description": "Updated location of the user",
                                        "type": "string",
                                        "example": "San Francisco"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Delete a user",
                "description": "Deletes a user by its ID.",
                "operationId": "e0f9a385a1d6788b99e519bd98cf7239",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID of the user to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Operation success status",
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "description": "Success message",
                                            "type": "string",
                                            "example": "User deleted successfully."
                                        },
                                        "data": {
                                            "description": "Additional response data",
                                            "type": "object",
                                            "example": null
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/users/{user}/assign-retailers": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Assign retailers to a user",
                "description": "Assigns a list of retailers to a specific user.",
                "operationId": "85c69a1b9c1923c7573701c91238ad2e",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID of the user to assign retailers to",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "retailers": {
                                        "description": "List of retailer IDs to assign",
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 10
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Retailers assigned successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/RetailerResource"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/users/{user}/revoke-retailers": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Revoke retailers from a user",
                "description": "Removes a list of retailers from a specific user.",
                "operationId": "682df37d765eda8379b5aff9485e1e69",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID of the user to revoke retailers from",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "retailers": {
                                        "description": "List of retailer IDs to revoke",
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 10
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Retailers revoked successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/RetailerResource"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "CurrencyResource": {
                "description": "Currency resource schema",
                "properties": {
                    "id": {
                        "description": "ID of the currency",
                        "type": "integer",
                        "example": 1
                    },
                    "code": {
                        "description": "Currency code",
                        "type": "string",
                        "example": "USD"
                    },
                    "name": {
                        "description": "Currency name",
                        "type": "string",
                        "example": "US Dollar"
                    },
                    "symbol": {
                        "description": "Currency symbol",
                        "type": "string",
                        "example": "$"
                    }
                },
                "type": "object"
            },
...
            "RetailerResource": {
                "description": "Retailer resource schema",
                "properties": {
                    "id": {
                        "description": "ID of the retailer",
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "description": "Title or name of the retailer",
                        "type": "string",
                        "example": "Best Buy"
                    },
                    "url": {
                        "description": "URL of the retailer",
                        "type": "string",
                        "example": "http://www.bestbuy.com"
                    },
                    "logo": {
                        "description": "Logo URL of the retailer",
                        "type": "string",
                        "example": "http://www.bestbuy.com/logo.png"
                    },
                    "currency": {
                        "$ref": "#/components/schemas/CurrencyResource"
                    }
                },
                "type": "object"
            },
...
            "UserResource": {
                "description": "User resource schema",
                "properties": {
                    "id": {
                        "description": "ID of the user",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Name of the user",
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "description": "Email address of the user",
                        "type": "string",
                        "example": "johndoe@example.com"
                    },
                    "is_verified": {
                        "description": "Whether the user's email is verified",
                        "type": "boolean",
                        "example": true
                    },
                    "role": {
                        "description": "Role of the user",
                        "type": "string",
                        "example": "REGULAR_USER"
                    },
                    "location": {
                        "description": "Location of the user",
                        "type": "string",
                        "example": "New York"
                    },
                    "retailers": {
                        "description": "List of retailers associated with the user",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/RetailerResource"
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}